requirements.txt
click==7.0
tabulate==0.8.6
pyYAML==5.1
pytest==4.1.0
pytest-cov==2.5.1
pytest-dependency==0.4.0
